Use Clojure core seqable? implementation when available
The previous implementation was flawed, and always used the library
defined version, even if the Clojure core version was availabe.

The check for the `seqable?` symbol didn't qualify it with the Clojure
core namespace, which resulted in the current (library) namespace
symbol being found (the very own `def` that was checking if it was
resolvable). Which in retrospect makes sense, as we were explicitly
excluding `seqable?` when refering Clojure core at the top.

We need to use the fully qualified name of the symbol to check for
Clojure core implementation. And if it is resolvable ("it exists"), we
need to `require` it before trying to use it. Otherwise we end up
using the current namespace symbol again.